requirements.txt
fastapi
uvicorn
pydantic
joblib
scikit-learn
google-genai